+++ src/sagemaker/session.py
@@ -673,6 +673,8 @@ class Session(object):  # pylint: disable=too-many-public-methods
         enable_network_isolation=None,
         image_uri=None,
         training_image_config=None,
+        container_entry_point=None,
+        container_arguments=None,
         algorithm_arn=None,
         encrypt_inter_container_traffic=None,
         use_spot_instances=False,
@@ -755,6 +757,11 @@ class Session(object):  # pylint: disable=too-many-public-methods
                 authenticate to the private Docker registry will be retrieved from this AWS Lambda
                 function. (default: ``None``). When it's set to None, SageMaker will not do
                 authentication before pulling the image in the private Docker registry.
+            container_entry_point (List[str]): Optional. The entrypoint script for a Docker
+                container used to run a training job. This script takes precedence over
+                the default train processing instructions.
+            container_arguments (List[str]): Optional. The arguments for a container used to run
+                a training job.
             algorithm_arn (str): Algorithm Arn from Marketplace.
             encrypt_inter_container_traffic (bool): Specifies whether traffic between training
                 containers is encrypted for the training job (default: ``False``).
@@ -855,6 +862,8 @@ class Session(object):  # pylint: disable=too-many-public-methods
             enable_network_isolation=enable_network_isolation,
             image_uri=image_uri,
             training_image_config=training_image_config,
+            container_entry_point=container_entry_point,
+            container_arguments=container_arguments,
             algorithm_arn=algorithm_arn,
             encrypt_inter_container_traffic=_encrypt_inter_container_traffic,
             use_spot_instances=use_spot_instances,
@@ -894,6 +903,8 @@ class Session(object):  # pylint: disable=too-many-public-methods
         enable_network_isolation=False,
         image_uri=None,
         training_image_config=None,
+        container_entry_point=None,
+        container_arguments=None,
         algorithm_arn=None,
         encrypt_inter_container_traffic=False,
         use_spot_instances=False,
@@ -976,6 +987,11 @@ class Session(object):  # pylint: disable=too-many-public-methods
                 authenticate to the private Docker registry will be retrieved from this AWS Lambda
                 function. (default: ``None``). When it's set to None, SageMaker will not do
                 authentication before pulling the image in the private Docker registry.
+            container_entry_point (List[str]): Optional. The entrypoint script for a Docker
+                container used to run a training job. This script takes precedence over
+                the default train processing instructions.
+            container_arguments (List[str]): Optional. The arguments for a container used to run
+                a training job.
             algorithm_arn (str): Algorithm Arn from Marketplace.
             encrypt_inter_container_traffic (bool): Specifies whether traffic between training
                 containers is encrypted for the training job (default: ``False``).
@@ -1042,6 +1058,12 @@ class Session(object):  # pylint: disable=too-many-public-methods
 
         if training_image_config is not None:
             train_request["AlgorithmSpecification"]["TrainingImageConfig"] = training_image_config
+
+        if container_entry_point is not None:
+            train_request["AlgorithmSpecification"]["ContainerEntrypoint"] = container_entry_point
+
+        if container_arguments is not None:
+            train_request["AlgorithmSpecification"]["ContainerArguments"] = container_arguments
 
         if algorithm_arn is not None:
             train_request["AlgorithmSpecification"]["AlgorithmName"] = algorithm_arn