+++ tests/unit/test_session.py
@@ -1657,6 +1657,8 @@ def test_train_with_sagemaker_config_injection(sagemaker_session):
             "TrainingRepositoryCredentialsProviderArn": "arn:aws:lambda:us-west-2:1234567897:function:test"
         },
     }
+    CONTAINER_ENTRY_POINT = ["bin/bash", "test.sh"]
+    CONTAINER_ARGUMENTS = ["--arg1", "value1", "--arg2", "value2"]
 
     sagemaker_session.train(
         image_uri=IMAGE,
@@ -1674,6 +1676,8 @@ def test_train_with_sagemaker_config_injection(sagemaker_session):
         enable_sagemaker_metrics=True,
         retry_strategy=RETRY_STRATEGY,
         training_image_config=TRAINING_IMAGE_CONFIG,
+        container_entry_point=CONTAINER_ENTRY_POINT,
+        container_arguments=CONTAINER_ARGUMENTS,
     )
 
     _, _, actual_train_args = sagemaker_session.sagemaker_client.method_calls[0]
@@ -1713,6 +1717,10 @@ def test_train_with_sagemaker_config_injection(sagemaker_session):
     assert (
         actual_train_args["AlgorithmSpecification"]["TrainingImageConfig"] == TRAINING_IMAGE_CONFIG
     )
+    assert (
+        actual_train_args["AlgorithmSpecification"]["ContainerEntrypoint"] == CONTAINER_ENTRY_POINT
+    )
+    assert actual_train_args["AlgorithmSpecification"]["ContainerArguments"] == CONTAINER_ARGUMENTS
     assert actual_train_args["RoleArn"] == expected_role_arn
     assert actual_train_args["ResourceConfig"] == {
         "InstanceCount": INSTANCE_COUNT,
@@ -1855,6 +1863,8 @@ def test_train_pack_to_request_with_optional_params(sagemaker_session):
             "TrainingRepositoryCredentialsProviderArn": "arn:aws:lambda:us-west-2:1234567897:function:test"
         },
     }
+    CONTAINER_ENTRY_POINT = ["bin/bash", "test.sh"]
+    CONTAINER_ARGUMENTS = ["--arg1", "value1", "--arg2", "value2"]
 
     sagemaker_session.train(
         image_uri=IMAGE,
@@ -1877,6 +1887,8 @@ def test_train_pack_to_request_with_optional_params(sagemaker_session):
         environment=ENV_INPUT,
         retry_strategy=RETRY_STRATEGY,
         training_image_config=TRAINING_IMAGE_CONFIG,
+        container_entry_point=CONTAINER_ENTRY_POINT,
+        container_arguments=CONTAINER_ARGUMENTS,
     )
 
     _, _, actual_train_args = sagemaker_session.sagemaker_client.method_calls[0]
@@ -1895,6 +1907,10 @@ def test_train_pack_to_request_with_optional_params(sagemaker_session):
     assert (
         actual_train_args["AlgorithmSpecification"]["TrainingImageConfig"] == TRAINING_IMAGE_CONFIG
     )
+    assert (
+        actual_train_args["AlgorithmSpecification"]["ContainerEntrypoint"] == CONTAINER_ENTRY_POINT
+    )
+    assert actual_train_args["AlgorithmSpecification"]["ContainerArguments"] == CONTAINER_ARGUMENTS
 
 
 def test_create_transform_job_with_sagemaker_config_injection(sagemaker_session):